refactor(srv): split rpc_types/memory.rs into bundle.rs + native_memory.rs
Phase 1 of SPEC_ARMORY_NAMING_CONSOLIDATION_2026_09_09.md. The file held both
concepts in one module (bundle command shapes + 19 NativeMemory* payloads).
Pure code motion: bundle.rs keeps the bundle types (names unchanged — two
generate frontend bindings and rename in the coordinated frontend PR),
native_memory.rs takes the NativeMemory* types (already correctly named).

No importer changes: rpc_types/mod.rs glob-re-exported memory::* and now
glob-re-exports both halves, so every rpc_types::X path resolves as before.
